backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.25
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-flask==1.2.0
google-genai>=1.60.0